get idRE

+++ src/equations.jl
@@ -38,7 +38,7 @@ function mme(f, userHints, userData, userPedData, blocks; paths2geno)
 	regionSizes = Array{Int64,1}(undef,0)
 		
 	#column id within pedigree
-	idRE = Array{Array{Float64,2},1}(undef,0)
+	idRE = Array{Array{Int64,2},1}(undef,0)
 	
         for i in 1:length(f.rhs)
 		if (f.rhs[i] isa FunctionTerm) && (String(nameof(f.rhs[i].forig)) == "PR")
